+++ R/TROLLv3_climatedaytime12.R
@@ -1,0 +1,18 @@
+#' TROLL climate parameters over months
+#'
+#' Climate parameters used by TROLL model over months.
+#'
+#' @format A data frame with 12 rows and 12 variables: \describe{
+#'   \item{Temperature}{temperature (degree C)}
+#'   \item{DaytimeMeanTemperature}{mean day temperature (degree C)}
+#'   \item{NightTemperature}{night temperature (degree C)}
+#'   \item{Rainfall}{rainfall (mm)} \item{WindSpeed}{wind speed (km/h)}
+#'   \item{DaytimeMeanIrradiance}{daityime mean irradiance (?)}
+#'   \item{MeanIrradiance}{mean irradiance (?)}
+#'   \item{SaturatedVapourPressure}{saturated vapour pressure (Pa)}
+#'   \item{VapourPressure}{vapour pressure (Pa)}
+#'   \item{VaporPressureDeficit}{vapour pressure deficit (VPD, Pa)}
+#'   \item{DayTimeVapourPressureDeficitVPDbasic}{daytime basic VPD (Pa)}
+#'   \item{DaytimeMeanVapourPressureDeficit}{daytime mean VPF (Pa)} }
+#'
+"TROLLv3_climatedaytime12"