+++ R/TROLLv3_climatedaytime365.R
@@ -1,0 +1,19 @@
+#' TROLL climate parameters over days
+#'
+#' Climate parameters used by TROLL model over days.
+#'
+#' @format A data frame with 365 rows and 12 variables: \describe{
+#'   \item{Temperature}{temperature (degree C)}
+#'   \item{DailyMeanTemperature}{mean day temperature (degree C)}
+#'   \item{NightTemperature}{night temperature (degree C)}
+#'   \item{Rainfall}{rainfall (mm)} \item{WindSpeed}{wind speed (km/h)}
+#'   \item{DailyMeanIrradiance}{daityime mean irradiance (?)}
+#'   \item{MeanIrradiance}{mean irradiance (?)}
+#'   \item{SaturatedVapourPressure}{saturated vapour pressure (Pa)}
+#'   \item{VapourPressure}{vapour pressure (Pa)}
+#'   \item{VaporPressureDeficit}{vapour pressure deficit (VPD, Pa)}
+#'   \item{DailyVapourPressureDeficit}{daytime basic VPD (Pa)}
+#'   \item{DailyMeanVapourPressureDeficit}{daytime mean VPF (Pa)} 
+#'   }
+#'
+"TROLLv3_climatedaytime365"