+++ R/TROLLv3_daytimevar.R
@@ -1,0 +1,10 @@
+#' TROLL daytime variation parameters
+#'
+#' Daytime variation parameters used by TROLL model.
+#'
+#' @format A data frame with 24 rows and 5 variables: \describe{
+#'   \item{starttime}{?} \item{endtime}{?} \item{vardaytime_light}{daytime
+#'   variation in light} \item{vardaytime_vpd}{daytime variation in vapor
+#'   pressure deficit} \item{vardaytime_T}{daytime variation in temperature}}
+#'
+"TROLLv3_daytimevar"