+++ R/TROLLv3_input.R
@@ -1,0 +1,9 @@
+#' TROLL global parameters
+#'
+#' Global parameters definition used by TROLL model.
+#'
+#' @format A data frame with 55 rows and 3 variables: \describe{
+#'   \item{param}{Global parameter} \item{value}{value of the parameter}
+#'   \item{description}{description of the parameter} }
+#'
+"TROLLv3_input"